Finished all error handling, validation

+++ src/main/java/com/demo/restapi/services/SellerService.java
@@ -2,6 +2,8 @@ package com.demo.restapi.services;
 
 import com.demo.restapi.dtos.SellerPreviewDTO;
 import com.demo.restapi.entities.Seller;
+import com.demo.restapi.exceptions.BusinessLogicException;
+import com.demo.restapi.exceptions.ResourceNotFoundException;
 import com.demo.restapi.mappers.SellerMapper;
 import com.demo.restapi.repositories.SellerRepository;
 import jakarta.transaction.Transactional;
@@ -27,11 +29,15 @@ public class SellerService {
     // 2. Get By ID
     public Seller getById(String id) {
         return sellerRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Seller not found"));
+                .orElseThrow(() -> new ResourceNotFoundException("Seller with id "+id+" was not found"));
     }
 
     // 3. Create
     public Seller createSeller(Seller seller) {
+        if (sellerRepository.existsByEmail(seller.getEmail())) {
+            // We throw our custom exception, NOT a generic RuntimeException
+            throw new BusinessLogicException("Email '" + seller.getEmail() + "' is already registered.");
+        }
         return sellerRepository.save(seller);
     }
 
@@ -49,6 +55,9 @@ public class SellerService {
 
     // 5. Delete
     public void deleteSeller(String id) {
+        if (!sellerRepository.existsById(id)){
+            throw new ResourceNotFoundException("Seller with id: "+id+" doesn't exist");
+        }
         sellerRepository.deleteById(id);
     }
